requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
pydantic==2.6.0
python-dotenv==1.0.1
requests==2.31.0
litellm>=1.81.0
openai>=1.12.0
